Switch to softer session removal.
The call to delete sessions when doing END SESSION handling was wrong.
In fact we need to be very careful about deleting sessions at all.  This
change prefers to mark sessions as closed instead of deleting them so
that backlogged messages are not corrupted.

Other work includes further enhancements to session data structures in
general along with some optimizations for getSessions to return the most
recently used sessions instead of the worst ones.  This speeds up
decryption for heavily used clients quite a bit.

Continued exception enhancements and moving the role of identity saving
and trust to higher level libs.  libsignal will no longer attempt to
save identity keys;  It merely asks for trust and punts if it doesn't
get the green light.